feat: move GenericLoadable, UrlPreviewState, TranslationConfig, ResourceToastMsg to commons
Move additional KMP-compatible utility classes to commons module:

- GenericLoadable<T>: generic sealed loading state (compose runtime only)
- UrlPreviewState: URL preview loading state (uses commons UrlInfoItem)
- TranslationConfig: translation configuration data class
- ResourceToastMsg: resource-based toast message

Updated 19 consumer files to import from commons where sealed class
nested types required direct imports. Fixed smart cast issues caused
by cross-module property access.
